==> R/dissolved_oxygen_uncorrected_mg_per_l/4.1_do_mg_per_l_spike.qmd <==
---
format: 
  html:
    toc: true
    toc-depth: 3
    embed-resources: true
editor_options: 
  chunk_output_type: console
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, dpi = 600, fig.width = 8)

library(dplyr)
library(DT)
library(here)
library(ggplot2)
library(leaflet)
library(lubridate)
library(plotly)
library(purrr)
library(qaqcmar)
library(RColorBrewer)
library(readr)
library(sensorstrings)
library(strings)
library(summaryplots)
library(tidyr)
library(viridis)

source(here("functions/filter_out_suspect_obs.R"))
source(here("functions/remove_do_correction.R"))
source(here("functions/subchunkify.R"))

theme_set(theme_light())

x_axis <- scale_x_continuous("Dissolved Oxygen (mg/L)")

dt_options <- list(
  dom = 'ft',
  paging = FALSE,
  searching = TRUE,
  scrollY = "500px",
  columnDefs = list(list(className = 'dt-center', targets = "_all"))
)

# data - preliminary QC 
dat_raw <- readRDS(here("data/do_mg_per_l_rolling_sd_prelim_qc.rds")) %>% 
  select(
    -c(sensor_type, int_sample, n_sample, 
       rolling_sd_flag_dissolved_oxygen_uncorrected_mg_per_l)) %>% 
  mutate(
    depth = factor(sensor_depth_at_low_tide_m),
    year = factor(year(timestamp_utc)),
    month = month(timestamp_utc)
  ) 

counties <- unique(dat_raw$county)

dat_all <- NULL

for(i in seq_along(counties)) {
  
  county_i <- counties[i]
  
  dat_all[[i]] <- dat_raw %>% 
    filter(county == county_i) %>% 
    qc_test_spike(keep_spike_cols = TRUE) %>% 
    select(-c(spike_high, spike_low, spike_flag_dissolved_oxygen_uncorrected_mg_per_l))
}

dat_all <- dat_all %>% 
  map_df(bind_rows) %>% 
  # filter_out_suspect_obs() and other foos expects data in the OLD format
  select(
    COUNTY = county, 
    STATION = station, 
    DEPLOYMENT_PERIOD = deployment_range,
    TIMESTAMP = timestamp_utc,
    VALUE = value_dissolved_oxygen_uncorrected_mg_per_l,
    DEPTH = sensor_depth_at_low_tide_m,
    everything()) %>% 
  mutate(
    VARIABLE = "Dissolved Oxygen", 
    UNITS = "mg/L", 
    YEAR = factor(year(TIMESTAMP))
  ) %>% 
  filter(!is.na(spike_value)) %>% 
  convert_depth_to_ordered_factor()

# filtered data
dat_filt <- dat_all %>% 
  filter_out_suspect_obs() 

county_pal <- get_county_colour_palette(length(unique(dat_all$COUNTY)))
depth_pal <- viridis(length(unique(dat_all$DEPTH)), direction = -1)
year_pal <- viridis( length(unique(dat_all$YEAR)), option = "C")

theme_facet_plotly <- theme(
  panel.spacing.y = unit(30, "lines"),
  panel.spacing.x = unit(20, "lines")
)

```

# Dissolved Oxygen (mg/L) Spikes

November 9, 2023

- Data submitted to the Open Data Portal in December 2022.
- Preliminary QC (submitted to Open Data Portal)
  - Obvious outliers removed.
  - Suspected biofouling removed.
- Filtered Data
  - Additional suspected biofouling was removed from the May 2022 deployments.

- Following QARTOD: $$spike_{value} = abs(value - spike_{ref})$$ where $spike_{ref} = (lead_{value} - lag_{value}) / 2$

## Preliminary QC 

### Pooled

::: panel-tabset

#### Figure 1: All observations
```{r}
#| label: fig-dist-all
#| fig-height: 6
#| fig-cap: Distribution of dissolved oxygen observations (binwidth = 0.1 mg/L).

p <- plot_histogram(dat_all, hist_col = spike_value, binwidth = 0.1) +
  x_axis

ggplotly(p)
```

#### Figure 2: By depth
```{r}
#| label: fig-dist-all-depth
#| fig-height: 6
#| fig-cap: Distribution of dissolved oxygen observations (binwidth = 0.1 mg/L), coloured by depth.

p <- plot_histogram(
  dat_all, hist_col = spike_value, binwidth = 0.1,
  fill_col = DEPTH, pal = depth_pal
) +
  x_axis

ggplotly(p)
```

#### Figure 3: By year
```{r}
#| label: fig-dist-all-year
#| fig-height: 6
#| fig-cap: Distribution of dissolved oxygen observations (binwidth = 0.1 mg/L), coloured by year.

p <- plot_histogram(
  dat_all, hist_col = spike_value, binwidth = 0.1, 
  fill_col = YEAR, pal = year_pal
) +
  x_axis

ggplotly(p)
```

:::

### County

::: panel-tabset

#### Figure 4: County 
```{r}
#| label: fig-dist-county
#| fig-height: 6
#| fig-cap: Distribution of dissolved oxygen observations by county (binwidth = 0.1 mg/L).

p <- dat_all %>% 
  plot_histogram(hist_col = spike_value, binwidth = 0.1) +
  x_axis +
  facet_wrap(~COUNTY, scales = "free", ncol = 3) +
  theme_facet_plotly

ggplotly(p)
```

#### Figure 5: County + Depth
```{r}
#| label: fig-dist-county-depth
#| fig-height: 6
#| fig-cap: Distribution of dissolved oxygen observations (binwidth = 0.1 mg/L), within each county, coloured by depth.

p <- plot_histogram(
  dat_all, hist_col = spike_value, binwidth = 0.1, fill_col = DEPTH, pal = depth_pal
) +
  x_axis +
  facet_wrap(~COUNTY, scales = "free", ncol = 3) +
  theme_facet_plotly
  
ggplotly(p)
```

#### Figure 6: County + Year
```{r}
#| label: fig-dist-county-year
#| fig-height: 6
#| fig-cap: Distribution of dissolved oxygen observations (binwidth = 0.1 mg/L), within each county coloured by year.

p <- plot_histogram(
  dat_all, hist_col = spike_value, binwidth = 0.1, fill_col = YEAR, pal = year_pal
) +
  x_axis +
  facet_wrap(~COUNTY, scales = "free", ncol = 3) +
  theme_facet_plotly

ggplotly(p)
```
:::

#### County + Month + Year

::: panel-tabset
```{r}
#| fig-height: 10
#| results: "asis"

counties <- unique(dat_all$COUNTY)

dat_all <- dat_all %>% 
  mutate(MONTH = month(TIMESTAMP, label = TRUE))

for(i in seq_along(counties)){
  
  county_i <- counties[i]
  dat_i <- dat_all %>% 
    filter(COUNTY == county_i) 
  
  cat('\n###', county_i, '\n')
  
  p <- plot_histogram(
    dat_i, hist_col = spike_value, binwidth = 0.1, fill_col = YEAR, pal = year_pal
  ) +
    x_axis +
    facet_wrap(~MONTH, ncol = 3)  
  
  print(p)
}

```
:::

### Summary Statistics

::: panel-tabset

#### Figure 7: Central Tendancy
```{r}
#| fig-height: 6
#| fig-cap: The mean and standard deviation of spike value in each county.
#| label: fig-mean-all

stats_all <- dat_all %>% 
  summarise(
    max = max(spike_value),
    mean = mean(spike_value),
    med = median(spike_value),
    stdev = sd(spike_value),
    
    q_90 = quantile(spike_value, prob = 0.90),
    q_95 = quantile(spike_value, prob = 0.95),
    q_997 = quantile(spike_value, prob = 0.997)
  ) %>% 
  mutate(COUNTY = "All") %>% 
  bind_rows(
    dat_all %>% 
      group_by(COUNTY) %>% 
      summarise(
        max = max(spike_value),
        mean = mean(spike_value),
        med = median(spike_value),
        stdev = sd(spike_value),
        
        q_90 = quantile(spike_value, prob = 0.90),
        q_95 = quantile(spike_value, prob = 0.95),
        q_997 = quantile(spike_value, prob = 0.997)
      ) 
  ) %>%
  mutate(
    variable = "dissolved_oxygen_uncorrected_mg_per_l",
    thresh_3sd = mean + 3 * stdev,
    across(.cols = where(is.numeric), .fns = round, digits = 2)
  ) %>% 
  select(county = COUNTY, everything())

p <- plot_mean_sd_county(stats_all)

ggplotly(p)
```

#### Figure 8: Quantiles
```{r}
#| fig-cap: Upper quantiles.
#| fig-height: 6
#| label: fig-quantiles-all

p <- stats_all %>% 
  select(county, contains("q")) %>% 
  pivot_longer(cols = contains("q"), names_to = "quantile") %>% 
  ggplot(aes(quantile, value, group = county, col = county)) +
  geom_point() +
  geom_line() +
  scale_color_manual(values = c("grey20", county_pal), drop = FALSE) +
  scale_y_continuous(limits = c(0, 2))

ggplotly(p)
```

#### Table 1: Summary statistics
```{r}
stats_all %>% 
  select(-c(variable, max, contains("thresh"))) %>% 
  datatable(
    options = list(
      dom = 'ft',
      paging = FALSE,
      searching = TRUE,
      columnDefs = list(list(className = 'dt-center', targets = "_all"))
    ), rownames = FALSE)
```
:::




## Filtered Data

### Pooled

::: panel-tabset

#### Figure 9: All observations
```{r}
#| fig-height: 6
#| fig-cap: Distribution of dissolved oxygen observations (binwidth = 0.1 mg/L).

p <- plot_histogram(dat_filt, hist_col = spike_value, binwidth = 0.1) +
  x_axis

ggplotly(p)
```

#### Figure 10: By depth
```{r}
#| fig-height: 6
#| fig-cap: Distribution of dissolved oxygen observations (binwidth = 0.1 mg/L), coloured by depth.

p <- plot_histogram(
  dat_filt, hist_col = spike_value, binwidth = 0.1,
  fill_col = DEPTH, pal = depth_pal
) +
  x_axis

ggplotly(p)
```

#### Figure 11: By year
```{r}
#| fig-height: 6
#| fig-cap: Distribution of dissolved oxygen observations (binwidth = 0.1 mg/L), coloured by year.

p <- plot_histogram(
  dat_all, hist_col = spike_value, binwidth = 0.1, 
  fill_col = YEAR, pal = year_pal
) +
  x_axis

ggplotly(p)
```

:::

### County

::: panel-tabset

#### Figure 12: County 
```{r}
#| fig-height: 6
#| fig-cap: Distribution of dissolved oxygen observations by county (binwidth = 0.1 mg/L).

p <- dat_filt %>% 
  plot_histogram(hist_col = spike_value, binwidth = 0.1) +
  x_axis +
  facet_wrap(~COUNTY, scales = "free", ncol = 3) +
  theme_facet_plotly

ggplotly(p)
```

#### Figure 13: County + Depth
```{r}
#| fig-height: 6
#| fig-cap: Distribution of dissolved oxygen observations (binwidth = 0.1 mg/L), within each county, coloured by depth.

p <- plot_histogram(
  dat_all, hist_col = spike_value, binwidth = 0.1, fill_col = DEPTH, pal = depth_pal
) +
  x_axis +
  facet_wrap(~COUNTY, scales = "free", ncol = 3) +
  theme_facet_plotly
  
ggplotly(p)
```

#### Figure 14: County + Year
```{r}
#| fig-height: 6
#| fig-cap: Distribution of dissolved oxygen observations (binwidth = 0.1 mg/L), within each county coloured by year.

p <- plot_histogram(
  dat_all, hist_col = spike_value, binwidth = 0.1, fill_col = YEAR, pal = year_pal
) +
  x_axis +
  facet_wrap(~COUNTY, scales = "free", ncol = 3) +
  theme_facet_plotly

ggplotly(p)
```
:::

#### County + Month + Year

::: panel-tabset
```{r}
#| fig-height: 10
#| results: "asis"

dat_filt <- dat_filt %>% 
  mutate(MONTH = month(TIMESTAMP, label = TRUE))

for(i in seq_along(counties)){
  
  county_i <- counties[i]
  dat_i <- dat_filt%>% 
    filter(COUNTY == county_i) 
  
  cat('\n###', county_i, '\n')
  
  p <- plot_histogram(
    dat_i, hist_col = spike_value, binwidth = 0.1, fill_col = YEAR, pal = year_pal
  ) +
    x_axis +
    facet_wrap(~MONTH, ncol = 3)  
  
  print(p)
}

```
:::

### Summary Statistics

::: panel-tabset

#### Figure 15: Central Tendancy
```{r}
#| fig-height: 6
#| fig-cap: The mean and standard deviation of spike value in each county.
#| label: fig-mean-filt

stats_filt <- dat_filt %>% 
  summarise(
    max = max(spike_value),
    mean = mean(spike_value),
    med = median(spike_value),
    stdev = sd(spike_value),
    
    q_90 = quantile(spike_value, prob = 0.90),
    q_95 = quantile(spike_value, prob = 0.95),
    q_997 = quantile(spike_value, prob = 0.997)
  ) %>% 
  mutate(COUNTY = "All") %>% 
  bind_rows(
    dat_filt %>% 
      group_by(COUNTY) %>% 
      summarise(
        max = max(spike_value),
        mean = mean(spike_value),
        med = median(spike_value),
        stdev = sd(spike_value),
        
        q_90 = quantile(spike_value, prob = 0.90),
        q_95 = quantile(spike_value, prob = 0.95),
        q_997 = quantile(spike_value, prob = 0.997)
      ) 
  ) %>%
  mutate(
    variable = "dissolved_oxygen_uncorrected_mg_per_l",
    thresh_3sd = mean + 3 * stdev,
    across(.cols = where(is.numeric), .fns = round, digits = 2)
  ) %>% 
  select(county = COUNTY, everything())

p <- plot_mean_sd_county(stats_filt)

ggplotly(p)
```

#### Figure 16: Quantiles
```{r}
#| fig-cap: Upper quantiles.
#| fig-height: 6
#| label: fig-quantiles-filt

p <- stats_filt %>% 
  select(county, contains("q")) %>% 
  pivot_longer(cols = contains("q"), names_to = "quantile") %>% 
  ggplot(aes(quantile, value, group = county, col = county)) +
  geom_point() +
  geom_line() +
  scale_color_manual(values = c("grey20", county_pal), drop = FALSE) +
  scale_y_continuous(limits = c(0, 2))

ggplotly(p)
```

#### Table 2: Summary statistics
```{r}
stats_filt %>% 
  select(-c(variable, max, contains("thresh"))) %>% 
  datatable(
    options = list(
      dom = 'ft',
      paging = FALSE,
      searching = TRUE,
      columnDefs = list(list(className = 'dt-center', targets = "_all"))
    ), rownames = FALSE)
```
:::
